api/assessment-api/requirements.txt
flask==3.0.0
flask-cors
python-dotenv==1.0.0
openai==1.3.7
jinja2==3.1.2
werkzeug==3.0.0
pathlib
